src/third_party/blink/web_tests/external/wpt/intersection-observer/isIntersecting-threshold-expected.txt

This is a testharness.js-based test.
PASS At initial scroll position
FAIL Scrolled to half way through target element assert_equals: expected true but got false
PASS Scrolled to target element completely off screen
Harness: the test ran to completion.
